archive/guest-talks/2018-django/retrosocial/requirements.txt
django
django-heroku
channels-redis
daphne
